requirements.txt
numpy
torch
torchvision
PILLOW
matplotlib
